idx Integer 형에서 Long 형으로 변경

+++ backend/src/main/java/com/example/backend/product/model/Product.java
@@ -20,7 +20,7 @@ import java.util.List;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer productIdx;
+    private Long productIdx;
     private String name;
     private double price;
     private String brand;
