+++ backend/src/main/java/com/example/backend/product/model/spec/SsdSpec.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class SsdSpec {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer ssdIdx;
+    private Long ssdIdx;
     private Integer ssdCapacity;
     private Integer ssdRead;
     private Integer ssdWrite;
